# src/CertManagerClient/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "CertManager API",
    "version": "v1"
  },
  "paths": {
    "/{organization-id}/api/v1/Certificate": {
      "post": {
        "tags": [
          "Certificate"
        ],
        "operationId": "CreateCertificate",
        "parameters": [
          {
            "name": "organization-id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CertificateModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CertificateModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CertificateModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CertificateModelWithId"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CertificateModelWithId"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CertificateModelWithId"
                }
              }
            }
          }
        }
      }
    },
    "/{organization-id}/api/v1/Certificates/{id}": {
      "get": {
        "tags": [
          "Certificate"
        ],
        "operationId": "GetCertificateById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "organization-id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CertificateModelWithId"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CertificateModelWithId"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CertificateModelWithId"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Certificate"
        ],
        "operationId": "DeleteCertificateById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "organization-id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Certificate"
        ],
        "operationId": "EditCertificateById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "organization-id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CertificateUpdateModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CertificateUpdateModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CertificateUpdateModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CertificateModelWithId"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CertificateModelWithId"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CertificateModelWithId"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/{organization-id}/api/v1/Certificates": {
      "get": {
        "tags": [
          "Certificate"
        ],
        "operationId": "GetAllCertificates",
        "parameters": [
          {
            "name": "TagsToSearch",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "TagsSearchBehavior",
            "in": "query",
            "style": "form",
            "schema": {
              "$ref": "#/components/schemas/CertificateSearchBehavior"
            }
          },
          {
            "name": "organization-id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CertificateModelWithId"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CertificateModelWithId"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CertificateModelWithId"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CertificateModel": {
        "type": "object",
        "properties": {
          "isCertificateAuthority": {
            "type": "boolean"
          },
          "certificateName": {
            "type": "string",
            "nullable": true
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "certificateDescription": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CertificateModelWithId": {
        "type": "object",
        "properties": {
          "isCertificateAuthority": {
            "type": "boolean"
          },
          "certificateName": {
            "type": "string",
            "nullable": true
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "certificateDescription": {
            "type": "string",
            "nullable": true
          },
          "certificateId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
...
      "CertificateSearchBehavior": {
        "enum": [
          "MatchAll",
          "MatchAny"
        ],
        "type": "string"
      },
      "CertificateUpdateModel": {
        "type": "object",
        "properties": {
          "newCertificateName": {
            "type": "string",
            "nullable": true
          },
          "newTags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "newCertificateDescription": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      }
    },
    "securitySchemes": {
      "JWTBearerAuth": {
        "type": "oauth2",
        "description": "Specify the authorization token.",
        "flows": {
          "password": {
            "authorizationUrl": "http://localhost:8080/realms/master/protocol/openid-connect/auth",
            "tokenUrl": "http://localhost:8080/realms/master/protocol/openid-connect/token",
            "scopes": {
              "cert-manager/read": "Read access",
              "cert-manager/write": "write access"
            }
          },
          "clientCredentials": {
            "authorizationUrl": "http://localhost:8080/realms/master/protocol/openid-connect/auth",
            "tokenUrl": "http://localhost:8080/realms/master/protocol/openid-connect/token",
            "scopes": {
              "cert-manager/read": "Read access",
              "cert-manager/write": "write access"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "JWTBearerAuth": [ ]
    }
  ]
}